BACKEND/requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
langchain==0.0.95
langchain-groq==0.0.2
langchain-text-splitter==0.0.3
langchain-chains==0.0.2
langchain-core==0.0.2
langchain-community==0.0.3
langchain-huggingface==0.0.1
faiss-cpu==1.7.2
